+++ cace/common/common.py
@@ -29,26 +29,14 @@ from ..logging import (
 
 def get_pdk(magicfilename=None):
     """
-    Get a value for the PDK, either from the second line of a .mag file,
-    or from the environment as environment variable "PDK".
-
-    NOTE:  Normally the PDK is provided as part of the datasheet, as
-    a project does not necessarily have a .mag file;  so there is no
-    source for automatically determining the project PDK.
+    Get a value for the PDK, as environment variable "PDK".
     """
-    if magicfilename and os.path.isfile(magicfilename):
-        with open(magicfilename, 'r') as ifile:
-            for line in ifile.readlines():
-                tokens = line.split()
-                if tokens[0] == 'tech':
-                    pdk = tokens[1]
-                    break
-    else:
-        try:
-            pdk = os.environ['PDK']
-        except KeyError:
-            error('No .mag file and PDK is not defined in the environment.')
-            pdk = None
+
+    try:
+        pdk = os.environ['PDK']
+    except KeyError:
+        error('PDK is not defined in the environment.')
+        pdk = None
 
     return pdk
 
@@ -78,12 +66,14 @@ def get_pdk_root():
         if pdk_root:
             os.environ['PDK_ROOT'] = pdk_root
         else:
-            error('Could not locate PDK_ROOT!')
+            error(
+                'PDK_ROOT is not defined in the environment and could not automatically locate PDK_ROOT.'
+            )
 
     return pdk_root
 
 
-def get_magic_rcfile(datasheet, magicfilename=None):
+def get_magic_rcfile():
     """
     Get the path and filename of the magic startup script corresponding
     to the PDK.
@@ -92,33 +82,55 @@ def get_magic_rcfile(datasheet, magicfilename=None):
     script (.magicrc file).
     """
 
-    if 'PDK_ROOT' in datasheet:
-        pdk_root = datasheet['PDK_ROOT']
-    else:
-        pdk_root = get_pdk_root()
-
-    if 'PDK' in datasheet:
-        pdk = datasheet['PDK']
-    elif magicfilename:
-        pdk = get_pdk(magicfilename)
-    else:
-        paths = datasheet['paths']
-        if magicfilename:
-            pdk = get_pdk(magicfilename)
-        elif 'magic' in paths:
-            magicpath = paths['magic']
-            magicfilename = os.path.join(magicpath, magicname)
-            pdk = get_pdk(magicfilename)
-        else:
-            return None
+    pdk_root = get_pdk_root()
+    pdk = get_pdk()
 
     rcfile = os.path.join(
         pdk_root, pdk, 'libs.tech', 'magic', pdk + '.magicrc'
     )
+
     return rcfile
 
 
-def get_netgen_setupfile(datasheet):
+def get_klayout_techfile():
+    """
+    Get the path and filename of the klayout tech file corresponding
+    to the PDK.
+
+    Returns a string containing the full path and filename of the tech
+    file (.lyt file).
+    """
+
+    pdk_root = get_pdk_root()
+    pdk = get_pdk()
+
+    techfile = os.path.join(
+        pdk_root, pdk, 'libs.tech', 'klayout', 'tech', pdk + '.lyt'
+    )
+
+    return techfile
+
+
+def get_klayout_layer_props():
+    """
+    Get the path and filename of the klayout layer properties corresponding
+    to the PDK.
+
+    Returns a string containing the full path and filename of the layer
+    properties (.lyp file).
+    """
+
+    pdk_root = get_pdk_root()
+    pdk = get_pdk()
+
+    techfile = os.path.join(
+        pdk_root, pdk, 'libs.tech', 'klayout', 'tech', pdk + '.lyp'
+    )
+
+    return techfile
+
+
+def get_netgen_setupfile():
     """
     Get the path and filename of the netgen setup script corresponding
     to the PDK.
@@ -127,27 +139,13 @@ def get_netgen_setupfile(datasheet):
     script (.tcl file).
     """
 
-    if 'PDK_ROOT' in datasheet:
-        pdk_root = datasheet['PDK_ROOT']
-    else:
-        pdk_root = get_pdk_root()
-
-    if 'PDK' in datasheet:
-        pdk = datasheet['PDK']
-    elif magicfilename:
-        pdk = get_pdk(magicfilename)
-    else:
-        paths = datasheet['paths']
-        if 'magic' in paths:
-            magicpath = paths['magic']
-            magicfilename = os.path.join(magicpath, magicname)
-            pdk = get_pdk(magicfilename)
-        else:
-            return None
+    pdk_root = get_pdk_root()
+    pdk = get_pdk()
 
     setupfile = os.path.join(
         pdk_root, pdk, 'libs.tech', 'netgen', pdk + '_setup.tcl'
     )
+
     return setupfile
 
 
@@ -235,6 +233,221 @@ def set_xschem_paths(datasheet, symbolpath, tclstr=None):
                     tcllist.append('append XSCHEM_LIBRARY_PATH :' + dependdir)
 
     return ' ; '.join(tcllist)
+
+
+def xschem_generate_svg(schempath, svgpath):
+    """
+    Generate an SVG drawing of a schematic or symbol using xschem
+
+    Return 0 if the drawing was generated, 1 if not.
+    """
+
+    if not os.path.isfile(schempath):
+        err(f'Could not find {schempath}.')
+        return 1
+
+    # Xschem arguments:
+    # -r:  Bypass readline (because stdin/stdout are piped)
+    # -x:  No X11 / No GUI window
+    # -q:  Quit after processing command line
+
+    xschemargs = [
+        '-r',
+        '-x',
+        '-q',
+        '--svg',
+        '--plotfile',
+        svgpath,
+    ]
+
+    pdk_root = get_pdk_root()
+    pdk = get_pdk()
+
+    # See if there is an xschemrc file  we can source
+    xschemrcfile = os.path.join(os.path.split(schempath)[0], 'xschemrc')
+    if os.path.isfile(xschemrcfile):
+        xschemargs.extend(['--rcfile', xschemrcfile])
+    else:
+        warn(f'No project xschemrc file found at: {xschemrcfile}')
+        warn(
+            f'It is highly recommended to set up an xschemrc file for your project.'
+        )
+
+        # Use the PDK xschemrc file for xschem startup
+        xschemrcfile = os.path.join(
+            pdk_root, pdk, 'libs.tech', 'xschem', 'xschemrc'
+        )
+        warn(f'Using the PDK xschemrc instead…')
+        if os.path.isfile(xschemrcfile):
+            xschemargs.extend(['--rcfile', xschemrcfile])
+        else:
+            err(f'No xschemrc file found in the {pdk} PDK!')
+
+    xschemargs.append(schempath)
+
+    dbg('Generating SVG using xschem.')
+
+    returncode = run_subprocess('xschem', xschemargs, write_file=False)
+
+    if returncode != 0:
+        return 1
+
+    return 0
+
+
+def magic_generate_svg(layout_path, svgpath):
+    """
+    Generate an SVG drawing of a layout using magic
+
+    Return 0 if the drawing was generated, 1 if not.
+    """
+
+    if not os.path.isfile(layout_path):
+        err(f'Could not find {layout_path}.')
+        return 1
+
+    layout_directory = os.path.split(layout_path)[0]
+    layout_filename = os.path.split(layout_path)[1]
+    layout_cellname = os.path.splitext(layout_filename)[0]
+    layout_extension = os.path.splitext(layout_filename)[1]
+
+    rcfile = get_magic_rcfile()
+
+    magic_input = ''
+
+    magic_input += f'addpath {os.path.abspath(layout_directory)}\n'
+    if layout_extension == '.mag':
+        magic_input += f'load {layout_filename}\n'
+    elif layout_extension == '.gds':
+        magic_input += f'gds read {layout_filename}\n'
+        magic_input += f'load {layout_cellname}\n'
+    else:
+        err(f'Unknown file extension for: {layout_path}')
+        return 1
+
+    magic_input += f'plot svg {svgpath}\n'
+
+    returncode = run_subprocess(
+        'magic',
+        ['-noconsole', '-d XR', '-rcfile', rcfile],
+        input=magic_input,
+        write_file=False,
+    )
+
+    if returncode != 0:
+        return 1
+
+    return 0
+
+
+def klayout_generate_png(layout_path, out_path):
+    """
+    Generate a PNG drawing of a layout using klayout
+
+    Return 0 if the drawing was generated, 1 if not.
+    """
+
+    if not os.path.isfile(layout_path):
+        err(f'Could not find {layout_path}.')
+        return 1
+
+    layout_directory = os.path.split(layout_path)[0]
+    layout_filename = os.path.split(layout_path)[1]
+    layout_cellname = os.path.splitext(layout_filename)[0]
+    layout_extension = os.path.splitext(layout_filename)[1]
+
+    techfile = get_klayout_techfile()
+    layer_props = get_klayout_layer_props()
+    pdk = get_pdk()
+
+    if pdk == 'sky130A':
+        tech_name = 'sky130'
+    elif pdk == 'sky130B':
+        tech_name = 'sky130'
+    else:
+        tech_name = pdk
+
+    klayout_script = """import pya
+import os
+
+# Input:
+# gds_path: path to the gds file
+# out_path: output directory
+# out_name: output name
+# w: width
+
+if not 'w' in globals():
+    w = 1024
+
+background_white = "#FFFFFF"
+background_black = "#000000"
+
+lv = pya.LayoutView()
+
+lv.set_config("grid-visible", "false")
+lv.set_config("grid-show-ruler", "false")
+lv.set_config("text-visible", "false")
+tech = pya.Technology.technology_by_name(tech_name)
+lv.load_layout(gds_path, tech.load_layout_options, tech_name)
+lv.max_hier()
+
+ly = lv.active_cellview().layout()
+
+# top cell bounding box in micrometer units
+bbox = ly.top_cell().dbbox()
+
+# compute an image size having the same aspect ratio than 
+# the bounding box
+h = int(0.5 + w * bbox.height() / bbox.width())
+
+lv.load_layer_props(layer_props)
+
+lv.set_config("background-color", background_white)
+lv.save_image_with_options(os.path.join(out_path, out_name + "_w.png"), w, h, 0, 0, 0, bbox, False)
+
+lv.set_config("background-color", background_black)
+lv.save_image_with_options(os.path.join(out_path, out_name + "_b.png"), w, h, 0, 0, 0, bbox, False)"""
+
+    scriptpath = 'klayout_script.py'
+
+    with open(scriptpath, 'w') as f:
+        f.write(klayout_script)
+
+    # -b: batch mode
+    # -nn: tech file
+    # -r: script
+    # -rd <name>=<value>: script variable
+
+    returncode = run_subprocess(
+        'klayout',
+        [
+            '-b',
+            '-nn',
+            techfile,
+            '-r',
+            scriptpath,
+            '-rd',
+            f'gds_path={layout_path}',
+            '-rd',
+            f'out_path={out_path}',
+            '-rd',
+            f'out_name={layout_cellname}',
+            '-rd',
+            f'tech_name={tech_name}',
+            '-rd',
+            f'layer_props={layer_props}',
+        ],
+        write_file=False,
+    )
+
+    # Delete script after use
+    if os.path.isfile(scriptpath):
+        os.remove(scriptpath)
+
+    if returncode != 0:
+        return 1
+
+    return 0
 
 
 # -----------------------------------------------------------------------
@@ -397,7 +610,12 @@ def get_condition_names_used(template):
     return (condlist, default_cond)
 
 
-def run_subprocess(proc, args=[], env=None, input=None, cwd=None):
+def run_subprocess(
+    proc, args=[], env=None, input=None, cwd=None, write_file=True
+):
+
+    if not cwd:
+        cwd = os.getcwd()
 
     dbg(
         f'Subprocess {proc} {" ".join(args)} at \'[repr.filename][link=file://{os.path.abspath(cwd)}]{os.path.relpath(cwd)}[/link][/repr.filename]\'…'
@@ -431,7 +649,7 @@ def run_subprocess(proc, args=[], env=None, input=None, cwd=None):
                 dbg(line.rstrip('\n'))
 
         # Write stderr to file
-        if stderr:
+        if stderr and write_file:
             with open(
                 f'{os.path.join(cwd, proc)}_stderr.out', 'w'
             ) as stderr_file:
@@ -444,7 +662,7 @@ def run_subprocess(proc, args=[], env=None, input=None, cwd=None):
                 dbg(line.rstrip())
 
         # Write stdout to file
-        if stdout:
+        if stdout and write_file:
             with open(
                 f'{os.path.join(cwd, proc)}_stdout.out', 'w'
             ) as stdout_file: